fix some valgrind warnings
git-svn-id: svn://svn.tuxfamily.org/svnroot/tesseract/main@1506 b412e182-2ed4-4078-95c4-27dc59f57fa9

+++ src/engine/renderlights.cpp
@@ -1669,7 +1669,7 @@ static inline bool htcmp(const lighttileslice &x, const lighttileslice &y)
     return x.tile->band == y.tile->band &&
            x.priority == y.priority &&
            x.numlights == y.numlights &&
-           (!x.numlights || !memcmp(&x.tile->lights[x.offset], &y.tile->lights[y.offset], x.numlights*sizeof(int)));
+           (!x.numlights || !memcmp(&x.tile->lights[x.offset], &y.tile->lights[y.offset], x.numlights*sizeof(ushort)));
 }
 
 vector<lightinfo> lights;
